Działajacy hibernate przyklad reprezentatywny

+++ mavenstart/src/main/java/io/github/mat3e/HelloService.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 public class HelloService {
     static final String FALLBACK_NAME="world";
-    static final Lang FALLBACK_LANG= new Lang(1L,"Hello", "en");
+    static final Lang FALLBACK_LANG= new Lang(1,"Hello", "en");
 
     private LangRepository repository;
 
@@ -22,8 +22,8 @@ public class HelloService {
 
     String prepareGreeting(String name, String lang)
     {
-        Long langId;
-       try{    langId =Optional.ofNullable(lang).map(Long::valueOf).orElse(FALLBACK_LANG.getId()); //zabezpieczenie lang przed byciem nullem
+        Integer langId;
+       try{    langId =Optional.ofNullable(lang).map(Integer::valueOf).orElse(FALLBACK_LANG.getId()); //zabezpieczenie lang przed byciem nullem
        }catch(NumberFormatException e){
            langId=FALLBACK_LANG.getId();
